Added dropout flag, bug fixes for args passed

+++ util.py
@@ -424,7 +424,7 @@ def get_train_parser():
                          help="Model to use for crnn part of fcn + crnn")
     parser.add_argument('--pretrained', type=str2bool, default=True,
                         help="Pretrained unet model for fcn-crnn")
-    parser.add_argument('--dropout', type=int, default=.5,
+    parser.add_argument('--dropout', type=float, default=.5,
                         help="Dropout probability to be used")
     parser.add_argument('--model_path', type=str)
     parser.add_argument('--eval_on_test', type=str2bool, default=False)
